+++ src/components/directory/directory.component.jsx
@@ -1,4 +1,4 @@
-import CategoryItem from '../category-item.component';
+import CategoryItem from '../category-item/category-item.component';
 import './directory.styles.scss'
 
 const Directory  = ({categories}) => {
