+++ src/components/directory/directory.styles.scss
